compiler: implement basic generator
Originally, I imagined the output of the compiler would be a program written in C, which is able to solve the constraints for a given number of each node.

While working on it, I decided that this introduces too much complexity - and that I'd rather work on a simpler version for now.

Currently, the generated code doesn't do any solving. It only validates inputs and allocates some memory for the values.