compiler: Bug fix
`enforce_multi_arc_constraints` was getting stuck in an infinite loop in cases where it was trying to constrain variables with no valid values. I've added a condition to the loop that will simply quit the function if such a situation occurs. This allows `solve` to report that the error has occoured.