compiler: Refactor INDEX binary operations into a PROPERTY_ACCESS expression
The idea is to front-load finding the placeholder index and property offsets to the parser and resolver, so `convert_expression` doesn't have to.

I'm not sure what would ultimately be the most memory efficient approach for this? I've done this for now as it seems the most natural thing to do, then it can be fixed later if it proves to be a problem.