compiler: Bug fix
When we weren't repeating the check to see if `instance_index[n]` had exceeded the length of the array - we ended up with memory errors where we tried to index for instances beyond the length of the instances array.

I think this change should not have any other effects besides fixing this bug - but I'm only partially confident about that!