compiler: Allow intermediate placeholders to use the same instance values
Previously, when evaluating arcs, we would never test for situations where multiple placeholders pointed to the same instance.

Now that we use placeholders to represent intermediately values that were not specified by the user, it makes sense to only apply this rule to placeholders that are defined by the user.